:white_check_mark: add test when converting hsl -> rgb

+++ lymui/hsl.c
@@ -27,7 +27,6 @@
     hsl->h = hue;
     hsl->s = roundDigit(getSaturation(min, max, _l) * 100, 10);
     hsl->l = roundDigit(_l * 100, 10);
-    
     free(rgb);
     
     return hsl;
@@ -110,7 +109,7 @@ Rgb *getRgbValueFromHsl(Hsl *hsl) {
     
     // choose the luminace formula
     float temp_lum = 0.0f;
-    if (round(_l)) {
+    if (roundf(_l)) {
         temp_lum = (_l + _s) - (_l * _s);
     } else {
         temp_lum = _l * (1.0f + _s);
